+++ src/PostgresPoolError.ts
@@ -1,9 +1,10 @@
-export class ErrorWithCode extends Error {
+export class PostgresPoolError extends Error {
   public code: string;
 
   public constructor(message: string, code: string) {
     super(message);
 
+    this.name = 'PostgresPoolError';
     this.code = code;
   }
 }